Modificando modulo de auth con google
