+++ backend/src/auth/auth.module.ts
@@ -1,15 +1,15 @@
 /* eslint-disable prettier/prettier */
 import { forwardRef, Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
+import { AuthService } from './service/auth.service';
+import { AuthController } from './controllers/auth.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from '../user/entities/user.entity';
 import { UserModule } from '../user/user.module';
 import jwtConfig from '../config/jwt.config';
 import { ConfigModule } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
-import { GoogleAuthenticationController } from './social/google-authentication.controller';
-import { GoogleAuthenticationService } from './social/providers/google-authentication.service';
+import { GoogleAuthenticationController } from './controllers/google-authentication.controller';
+import { GoogleAuthenticationService } from './service/google-authentication.service';
 import { GenerateTokensProvider } from './providers/generate-tokens.provider';
 import { MailsModule } from 'src/mails/mails.module';
 import { AccessTokenGuard } from './guards/access-token.guard/access-token.guard';
