+++ backend/src/auth/dtos/google-authentication.dto.ts
@@ -1,3 +1,4 @@
+/* eslint-disable prettier/prettier */
 import { IsNotEmpty } from 'class-validator';
 
 export class GoogleTokenDto {
