+++ backend/src/liquidations/liquidations.service.ts
@@ -1,0 +1,10 @@
+/* eslint-disable prettier/prettier */
+import { Injectable } from '@nestjs/common';
+
+
+@Injectable()
+export class LiquidationsService {
+  constructor(
+  ) {}
+  
+}
