--- backend/src/payments_aux/payments.controller.ts
@@ -1,12 +1,0 @@
-/* eslint-disable prettier/prettier */
-import { Body, Controller, Post, Req } from '@nestjs/common';
-import { PaymentsService } from './payments.service';
-
-@Controller('payments')
-export class PaymentsController {
-  constructor(private readonly paymentService: PaymentsService) {}
-  @Post()
-  async generatePayment(@Req() request: Request, @Body() body: any) {
-    await this.paymentService.paymentLogic(request, body);
-  }
-}
