--- backend/src/payments_aux/payments.module.ts
@@ -1,14 +1,0 @@
-/* eslint-disable prettier/prettier */
-import { Module } from '@nestjs/common';
-import { PaymentsController } from './payments.controller';
-import { PaymentsService } from './payments.service';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from 'src/user/entities/user.entity';
-import { Payment } from './entities/payments.entity';
-
-@Module({
-  controllers: [PaymentsController],
-  providers: [PaymentsService],
-  imports: [TypeOrmModule.forFeature([User, Payment])],
-})
-export class PaymentsModule {}
