+++ backend/src/user/interfaces/email-user.interface.ts
@@ -1,3 +1,4 @@
+/* eslint-disable prettier/prettier */
 export interface EmailUser {
   name: string;
   email: string;
