+++ backend/src/user/interfaces/google-user.interface.ts
@@ -1,3 +1,4 @@
+/* eslint-disable prettier/prettier */
 export interface GoogleUser {
   email: string;
   name: string;
